added conditions for removing columns in remove unneeded files in doc tracking

+++ database/migrations/2016_11_28_074354_remove_unneeded_fields_in_doc_tracking.php
@@ -17,8 +17,13 @@ class RemoveUnneededFieldsInDocTracking extends Migration {
             DB::statement('TRUNCATE "tblCOM_DocTracking" cascade;');
 
             Schema::table('tblCOM_DocTracking', function ($table) {
-                $table->dropColumn('DT_FK_NSCode');
-                $table->dropColumn('DT_Amount');
+                if (Schema::hasColumn('tblCOM_DocTracking', 'DT_FK_NSCode')) {
+                    $table->dropColumn('DT_FK_NSCode');
+                }
+
+                if (Schema::hasColumn('tblCOM_DocTracking', 'DT_Amount')) {
+                    $table->dropColumn('DT_Amount');
+                }
             });
 
             DB::commit();
